Move compareDates.js tests to their own file
